+++ lib/src/utils/argon2.dart
@@ -1,16 +1,30 @@
 import 'dart:typed_data';
 
-import 'package:pointycastle/key_derivators/api.dart' show Argon2Parameters;
-import 'package:pointycastle/key_derivators/argon2.dart';
+import 'package:native_crypto/native_crypto.dart' as native_crypto;
 
 import '../bip39_kdf.dart';
 
+native_crypto.Argon2Version _mapArgon2Version(int version) {
+  switch (version) {
+    case Bip39Argon2Version.v10:
+      return native_crypto.Argon2Version.v10;
+    case Bip39Argon2Version.v13:
+      return native_crypto.Argon2Version.v13;
+    default:
+      throw ArgumentError.value(
+        version,
+        'version',
+        'unsupported Argon2 version: expected 0x10 (v1.0) or 0x13 (v1.3)',
+      );
+  }
+}
+
 /// Argon2id seed derivation (64-byte output by default).
 class Argon2 {
-  Argon2({Argon2BytesGenerator? generator})
-      : _generator = generator ?? Argon2BytesGenerator();
+  Argon2({native_crypto.Argon2? generator})
+      : _generator = generator ?? native_crypto.Argon2();
 
-  final Argon2BytesGenerator _generator;
+  final native_crypto.Argon2 _generator;
 
   static final Argon2 instance = Argon2();
 
@@ -19,16 +33,15 @@ class Argon2 {
     Uint8List salt, {
     Bip39Argon2Params params = Bip39Argon2Params.defaults,
   }) {
-    final argon2Params = Argon2Parameters(
-      Argon2Parameters.ARGON2_id,
-      salt,
-      desiredKeyLength: params.desiredKeyLength,
+    return _generator.deriveKey(
+      password: password,
+      salt: salt,
+      type: native_crypto.Argon2Type.argon2id,
+      version: _mapArgon2Version(params.version),
+      memoryKiB: params.memoryKiB,
       iterations: params.iterations,
-      memory: params.memoryKiB,
-      lanes: params.parallelism,
-      version: params.version,
+      parallelism: params.parallelism,
+      hashLength: params.desiredKeyLength,
     );
-    _generator.init(argon2Params);
-    return _generator.process(password);
   }
 }